requirements.txt
asttokens==2.4.1
decorator==5.1.1
executing==2.0.1
ipython==8.26.0
jedi==0.19.1
lxml==5.3.0
matplotlib-inline==0.1.7
numpy==1.26.4
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
wcwidth==0.2.13
